chore: streamline profile tests
