+++ tests/Bridge/Github/GithubProfileTest.php
@@ -13,7 +13,6 @@ declare(strict_types=1);
 
 namespace Sigwin\Ariadne\Test\Bridge\Github;
 
-use Nyholm\Psr7\Response;
 use Psr\Cache\CacheItemPoolInterface;
 use Psr\Http\Client\ClientInterface;
 use Psr\Http\Message\RequestInterface;
@@ -75,9 +74,6 @@ final class GithubProfileTest extends ProfileTestCase
         static::assertCount(1, $profile->getSummary()->getTemplates());
     }
 
-    /**
-     * @return iterable<array{string, bool|string}>
-     */
     protected function provideValidAttributeValues(): iterable
     {
         return [
@@ -91,9 +87,6 @@ final class GithubProfileTest extends ProfileTestCase
         ];
     }
 
-    /**
-     * @return iterable<array{string, int|bool|string}>
-     */
     protected function provideInvalidAttributeValues(): iterable
     {
         return [
@@ -103,28 +96,9 @@ final class GithubProfileTest extends ProfileTestCase
         ];
     }
 
-    /**
-     * @param array<string, string> $requests
-     */
-    protected function createHttpClient(array $requests = []): ClientInterface
+    protected function validateRequest(RequestInterface $request): void
     {
-        $httpClient = $this->getMockBuilder(ClientInterface::class)->getMock();
-
-        foreach ($requests as $url => $response) {
-            $httpClient
-                ->expects(static::once())
-                ->method('sendRequest')
-                ->willReturnCallback(static function (RequestInterface $request) use ($url, $response): Response {
-                    self::assertSame('GET', $request->getMethod());
-                    self::assertSame($url, $request->getUri()->__toString());
-                    self::assertSame('token ABC', $request->getHeaderLine('Authorization'));
-
-                    return new Response(200, ['Content-Type' => 'application/json'], $response);
-                })
-            ;
-        }
-
-        return $httpClient;
+        static::assertSame('token ABC', $request->getHeaderLine('Authorization'));
     }
 
     protected function createProfileInstance(ProfileConfig $config, ClientInterface $client, ProfileTemplateFactory $factory, CacheItemPoolInterface $cachePool): Profile
